--- a/src/inference/src/dev/threading/thread_affinity.hpp
+++ b/src/inference/src/dev/threading/thread_affinity.hpp
@@ -101,5 +101,13 @@ bool pin_current_thread_by_mask(int ncores, const CpuSet& processMask);
  * @return     `True` in case of success, `false` otherwise
  */
 bool pin_current_thread_to_socket(int socket);
+
+/**
+ * @brief      Set all cpus available
+ * @ingroup    ov_dev_api_threading
+ *
+ * @param[in]  ncpus  The number of cpus
+ */
+void set_all_cpus_available(int ncpus);
 }  // namespace threading
 }  // namespace ov