### Details:
 - *Fix the issue of cpu affinity mask changed on a NUMA machine with SEQ*

### Tickets:
 - *143272*