@dmitry-gorokhov 2024-07-11
General comment:
1. I don't think restoring affinity mask on compiled model destructor stage is a good enough solution. User application usually interleaves some additional computational logic with OV inference, so the pipeline looks in the following way: [user code] -> [OV infer] -> [user code] -> [OV Infer] -> .... -> [compiled model desctructor]. So the current solution still affects used code behavior.
I could propose 2 solution: a) Restore the mask at the end of Infer call b) Create separate thread for compilation/inference inside OV if enable pinning is ON.