@wangleis 2024-07-11
@dmitry-gorokhov Since CPU mask affects whole process, option b may not work. Xiaoxia will update PR to change and restore CPU mask during inferring like option a.