@sunxiaoxia2022 2024-07-15
@dmitry-gorokhov I put cpu pinning before and after the task(). Please take a look again.