@dmitry-gorokhov 2024-07-15
AS we discussed on the sync. Lets clarify 2 things first: 
1. What overheads is created by `[un]pin_stream_to_cpus()` calls on each inference?
2. Does OV created separate thread for inference in SEQ mode? If yes does it mean application thread is not affected by OV pinning/affinity behavior?