@sunxiaoxia2022 2024-07-16
> AS we discussed on the sync. Lets clarify 2 things first:
> 
> 1. What overheads is created by `[un]pin_stream_to_cpus()` calls on each inference?
> 2. Does OV created separate thread for inference in SEQ mode? If yes does it mean application thread is not affected by OV pinning/affinity behavior?

1. I tested the performance of master and this PR with a small model ebgan(onnx, FP32) by 10 times. 
Test machine: Intel(R) Xeon(R) Gold 6330 CPU @ 2.00GHz
Test command: ./benchmark_app -m /home/openvino-ci-69/xiaoxia/models/cv_bench_cache/ww21_weekly_23.0.0-10926-b4452d56304-API2.0/ebgan/onnx/onnx/FP32/1/dldt/ebgan.xml -d CPU -hint none -nstreams 1 -nthreads 1
Test result: the values are the average of 10 times. 
       master:  latency: 1.829 ms, throughput: 545.311 FPS
       This PR: latency: 1.833 ms, throughput: 543.561 FPS. latency increased 0.21% compared with master, throughput dropped 0.32%.

2. For asnyc mode, OV created separate thread for inference in SEQ which is the same as TBB. So application thread is not affected by OV pinning.
    But for sync mode, OV not create separate thread, it uses app thread. So app thread is affected by OV pinning.