@dmitry-gorokhov 2024-07-17
Thanks @sunxiaoxia2022 !
So to complete the task we need to fix the behavior for sync mode as well. I would propose to spawn separate thread in case of SEQ build, sync API amd enable_pinning==true.

@wangleis Do you have any objections?