@dmitry-gorokhov 2024-07-18
> @dmitry-gorokhov In sync mode, app is blocked during inference and app thread is used for inference. Since we will restore CPU mask after inference, current solution should be ok for sync mode from my understanding.

Good point. Agree. Approved!